refactor: Refactor class with constraint constants.
